Relocate case assignment delete API to cases rather than queues
